\name{soundecology}

\alias{soundecology}

\docType{package}

\title{Soundscape ecology}

\description{Functions to calculate indices for soundscape ecology and other ecology research that uses audio recordings. Still working on the algorithms, not all have been tested.}

\details{
	\tabular{ll}{
		Package: \tab soundecology\cr
		Type: \tab Package\cr
		Version: \tab 0.1alpha\cr
		Date: \tab 2013-06-29\cr
		License: \tab GPLv3\cr
		}
	}
\author{
Luis J. Villanueva-Rivera
}

\keyword{ soundecology }
