\name{soundecology}

\alias{soundecology}

\docType{package}

\title{Soundscape ecology}

\description{Functions to calculate indices for soundscape ecology and other ecology research that uses audio recordings.}

\details{
	\tabular{ll}{
		Package: \tab soundecology\cr
		Type: \tab Package\cr
		Version: \tab 0.42\cr
		Date: \tab 2013-10-18\cr
		License: \tab GPLv3\cr
		}
	}
\author{
Luis J. Villanueva-Rivera and Bryan C. Pijanowski
}

\keyword{ soundecology }
