\name{tropicalsound}

\alias{tropicalsound}

\title{tropicalsound sound example}

\description{Sample sound data of a recording of a tropical chorus of frogs.}

\usage{
data(tropicalsound)
}

\format{A Wave object.}

\details{
  Duration = 20 sec. Sampling rate = 22050 Hz.
}

\source{
  Recording made at a tropical rainforest in Puerto Rico by Luis J. Villanueva-Rivera.
}

\keyword{soundecol}
\keyword{tropicalsound}