\name{soundecol-package}
\alias{soundecol-package}
\alias{soundecol}
\docType{package}
\title{Functions for soundscape ecology and other research in ecology that uses audio recordings}
\description{Functions for soundscape ecology and other research in ecology that uses audio recordings. Check each function for the paper where it comes from.}
\details{
\tabular{ll}{
Package: \tab soundecol\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-06-27\cr
License: \tab GPLv3\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Luis J. Villanueva-Rivera

Maintainer: Luis J. Villanueva-Rivera <ljvillanueva@coquipr.com>
}
\references{Villanueva-Rivera, L. J., B. C. Pijanowski, J. Doucette, and B. Pekin. 2011. A primer of acoustic analysis for landscape ecologists. Landscape Ecology 26: 1233-1246. doi:10.1007/s10980-011-9636-9.}
\keyword{ package }
\examples{
}
