\name{soundecol}

\alias{soundecol}

\docType{package}

\title{Soundscape ecology}

\description{Indices for soundscape ecology and other research in ecology that uses audio recordings. Check each function for the reference paper for each index.}

\details{
	\tabular{ll}{
		Package: \tab soundecol\cr
		Type: \tab Package\cr
		Version: \tab 0.1\cr
		Date: \tab 2013-06-29\cr
		License: \tab GPLv3\cr
		}
	}
\author{
Luis J. Villanueva-Rivera
}

\keyword{ soundecol }