\name{soundecology}

\alias{soundecology}

\docType{package}

\title{Soundscape Ecology}

\description{Functions to calculate indices for soundscape ecology and other ecology research that uses audio recordings.}

\details{
	\tabular{ll}{
		Package: \tab soundecology\cr
		Type: \tab Package\cr
		Version: \tab 1.2\cr
		Date: \tab 2015-01-11\cr
		License: \tab GPLv3\cr
		}
	}
\author{
Luis J. Villanueva-Rivera and Bryan C. Pijanowski
}

\keyword{ soundecology }
