\name{tropicalsound}

\alias{tropicalsound}

\title{tropicalsound sound example}

\description{Sample sound of a digital recording of a chorus of tropical frogs.}

\usage{
data(tropicalsound)
}

\format{A \code{Wave} object.}

\details{
  Duration = 20 sec. Sampling rate = 22050 Hz.
}

\source{
  Recording made at a tropical rainforest in Puerto Rico by Luis J. Villanueva-Rivera.
}

\examples{
	data(tropicalsound)
	
	tropicalsound
	}
	
\keyword{soundecology}
\keyword{tropicalsound}
